+++ spec/helpers/student_loans_helper_spec.rb
@@ -39,7 +39,7 @@ describe StudentLoansHelper do
   describe "#student_loan_amount_question" do
     it "returns the question for the student laon amount question in the Student Loans journey, based on the configured current academic year" do
       expect(policy_configurations(:student_loans).current_academic_year).to eq AcademicYear.new("2025/2026")
-      expect(helper.student_loan_amount_question).to eq "Exactly how much student loan did you repay while employed as a teacher between 6 April 2024 and 5 April 2025?"
+      expect(helper.student_loan_amount_question).to eq "Exactly how much student loan did you repay in the 2024/2025 financial year?"
     end
   end
 end